moved new product form into its own component
